+++ R/which.pmin.R
@@ -9,6 +9,11 @@
 #' @details See original idea https://github.com/radiant-rstats/radiant.data/blob/master/R/radiant.R
 #' @name whichp
 #' @family whichp
+#' @examples
+#'
+#' which.pmin(a=1:5,b =5:1,silent = TRUE)
+#' which.pmin_chr(a=1:5,b =5:1,ties="first")
+#' which.pmin_chr(a=1:5,b =5:1,ties="collapse")